requirements.txt
black
langchain==0.1.5
langchain-openai==0.0.5
pydantic==2.6.1
python-dotenv==1.0.1
